Create 'role' package to manage role linking
This uses the 'acro' package to define each role as an acronym and
associate extra information with each role. Ideally the extra
information would be reproduced in the relevant section and also in the
Roles appendix. This requires version >=2 of acro, which I don't have
right now.

Each mention of a role links back to its definition section.