Create more custom role data management
Trying to hack an acronym/glossary handling package to store the email
address and telephone number for each role was just not going to work.
Also, as I began populating the list of roles, it became apparent that
referring to them via some acronym will get rather confusing. It's
probably much better to just always refer to them by the full title in
the document (in the source we can use acronyms)