Extract role title and contact info from dot files
There is now one source of each role title and we can maintain the
contact details in the same place. A python script extracts the relevant
data from the dot files and generates a CSV suitable for use with the
latex datatool package. There is the option to exclude certain roles
defined in the dot files from the CSV (specifically for the case where a
particular role is defined twice to link the charts). There is also an
option to exclude certain roles from the contacts table in the appendix
as not all roles will have generic contact details ('volunteers' roles
for example).