Add sectioning command arg to roledesc env
We have role descriptions in two parts of the document and they require
sections at different levels, hence the configurable command. Also, use
the starred versions of the sectioning commands to removing the
numbering and entries in the index. (The main reason I've done this is
that the role id is showing up as the index entry and I can't be
bothered to try to fix it right now).