Remove subordinates of Production Manager
These were both unnecessary (third parties) and too prescriptive. The
arena and pit wranglers are really just part of the group of people
'building' the competition and one of the responsibilities of the
production manager will be to ensure this happens, hence they are free
to decide how it happens. See #49.