Remove local {kickstart,activity} volunteers
These aren't really roles in themselves. It's a responsibility of the
Local Kickstart Coordinator and the Local Activity Coordinator to ensure
that there are sufficient people to help out, but there's no need for
these people to be explicitly shown in the org-chart.

Note that I've left mentors, even though the same argument could hold for
them, because they're so core to our organisation I feel that they are
worth showing.