Attempt to tie down kit roles
This splits the kit roles into hardware/software. Since the hardware
side involves things in the physical world the development process is
quite separate from the manufacture/maintenance process, unlike the
software side. The software side has a much faster cycle time between
development and deployment therefore it doesn't make much sense to break
it down past the 'Kit Software Coordinator' as they will define their
own unnamed roles as appropriate.

The big omission at the moment is that of a 'Support' role. I'm not sure
of we actually require that as a role in the org-chart. For now I plan
on adding support as a responsibility of the Kit Coordinator and they
are then free to create an unnamed role as appropriate to manage it.