Remove Kit Local Software proposal
This doesn't really fit with how software is developed. Hardware has a
very long cycle time and a large cost associated with the 'release'
(i.e. manufacturing). Software on the other hand has a very short cycle
time and releasing is free. Also, software is generally released with
few featured and extended in later releases.

We still need to keep track of any potential legal implications so make
it a requirement to maintain this list for each piece of software. This
list gets reviewed prior to each release.